-- Breaking the build (further) with some alterations to MovieID_Ratings files that will eventually reduce the on-disk footprint of Neustore

+++ MovieID_Ratings/UserRating.java
@@ -5,14 +5,14 @@ package MovieID_Ratings;
  */
 public class UserRating {
 	
-	public UserRating(int userId, int rating, String date) {
+	public UserRating(int userId, byte rating, String date) {
 		this.userId = userId;
 		this.rating = rating;
 		// TODO: this is probably not what we want
 		this.date = 0; //(int) Date.valueOf(date).getTime();
 	}
 	
-	public UserRating(int userId, int rating, int date) {
+	public UserRating(int userId, byte rating, int date) {
 		this.userId = userId;
 		this.rating = rating;
 		this.date = date;
@@ -23,7 +23,7 @@ public class UserRating {
 	}
 	
 	public final int userId;
-	public final int rating;
+	public final byte rating;
 	public final int date;
 
 }
